Vector --> Matrix Shaping Example
To do (spatial matrix) * (spacetime vector) it's useful to think of the spacetime vector as a matrix with a space index and a time index.  Then, multiply the spatial matrix against the space index, wind up with a matrix with again a space index and a time index, and interpret the result as a spacetime vector.

This, surprisingly, actually works in BLAZE!