Hotfix: irrep states are now diagonal.
numpy.linalg.eig and .eigh don't guarantee the orthogonality of the eigenvectors.  The issue is that those routines can find right eigenvalues for matrices that aren't necessarily hermitian, so there's no guarantee that eigenvectors *are* orthogonal!

However, to transform from the position basis to the basis of good irrep, one needs a unitary transformation.  If you pick irrep states that aren't orthogonal, you wind up with spurious overlap between eigenstates.

Fortunately, the fix is simple: use scipy.linalg.orth instead.  That routine provides an orthonormal basis.

CAVEAT USER!
============

Just because it provides an orthonormal basis for the hopping matrix doesn't mean all the states are necessarily the ones you might pick if you were doing a particular example by hand.  For example, the states it produces for the triangle are states of a parity across a particular reflection.  Were I doing it I'd pick states of good rotation-by-120˚.