I should learn to use git correctly.
In an attempt to remember how I diagonalized correlators in the past,
I checked out an old script (scripts/correlators.py) from a commit when isle was still called cns.

That brought the script into the future but snuck into the subsequent commit without a `git add`,
which I guess is the usual behavior for such a file.

This commit deletes that script, and the functionality belongs in archipelago anyway.

ANYway, I should learn to use git correctly.